development_widget add + ScoutingView and formationView revamp
